+++ docs/CHANGELOG.md
@@ -12,10 +12,13 @@ and here
 
 ### Changed
 -->
-## Release 1.3.16 -- 2023-05-07
+## Release 1.3.21 -- 2023-05-16
+-Fixed Issue Exclusive Gateway Convergence #87
+## Release 1.3.18 -- 2023-05-07
 - Fixed Issue # 84 https://github.com/ralphhanna/bpmn-server/issues/84
-  
-- item.data is now saved to the DB
+- Fixed Issue with Input Parameters #85
+- Fixed Request to update dependencies #79   
+
 ## Release 1.3.15 -- 2022-11-20
 
 - Added to .env